+++ src/Gadfly.jl
@@ -41,11 +41,6 @@ export Plot, Layer, Theme, Col, Row, Scale, Coord, Geom, Guide, Stat, Shape, ren
 export SVGJS, SVG, PGF, PNG, PS, PDF, draw, inch, mm, cm, px, pt, color, @colorant_str, vstack, hstack, title, gridstack
 
 
-function link_terminalextensions()
-    @debug "Loading TerminalExtensions support into Gadfly"
-    include("terminalextensions.jl")
-end
-
 function __init__()
     # Define an XML namespace for custom attributes
     Compose.xmlns["gadfly"] = "http://www.gadflyjl.org/ns"
@@ -63,8 +58,8 @@ function __init__()
 
     insert!(Base.Multimedia.displays, findlast(x->(x isa TextDisplay || x isa REPL.REPLDisplay), Base.Multimedia.displays)+1, GadflyDisplay())
 
-    @require DataFrames="a93c6f00-e57d-5684-b7b6-d8193f3e46c0" link_dataframes()
-    @require TerminalExtensions="d3a6a179-465e-5219-bd3e-0137f7fd17c7" link_terminalextensions()
+    @require DataFrames="a93c6f00-e57d-5684-b7b6-d8193f3e46c0" include("dataframes.jl")
+    @require TerminalExtensions="d3a6a179-465e-5219-bd3e-0137f7fd17c7" include("terminalextensions.jl")
 end
 
 
@@ -1114,7 +1109,7 @@ const default_aes_scales = Dict{Symbol, Dict}(
     ),
 
     :functional => Dict{Symbol, Any}(
-        :x      => Scale.x_discrete(),
+        :x      => Scale.x_func(),
         :z      => Scale.z_func(),
         :y      => Scale.y_func(),
         :shape  => Scale.shape_identity(),