+++ src/statistics.jl
@@ -513,18 +513,53 @@ is controlled by `bandwidth`.  Calls [`Stat.contour`](@ref) to compute the
 const density2d = Density2DStatistic
 
 function apply_statistic(stat::Density2DStatistic,
-                         scales::Dict{Symbol, Gadfly.ScaleElement},
-                         coord::Gadfly.CoordinateElement,
-                         aes::Gadfly.Aesthetics)
-    Gadfly.assert_aesthetics_defined("Density2DStatistic", aes, :x, :y)
+    scales::Dict{Symbol, Gadfly.ScaleElement},
+    coord::Gadfly.CoordinateElement,
+    aes::Gadfly.Aesthetics)
 
     window = (stat.bw[1] <= 0.0 ? KernelDensity.default_bandwidth(aes.x) : stat.bw[1],
               stat.bw[2] <= 0.0 ? KernelDensity.default_bandwidth(aes.y) : stat.bw[2])
-    k = KernelDensity.kde((aes.x,aes.y), bandwidth=window, npoints=stat.n)
-    aes.z = k.density
-    aes.x = collect(k.x)
-    aes.y = collect(k.y)
-    apply_statistic(ContourStatistic(levels=stat.levels), scales, coord, aes)
+
+    Dat = [aes.x aes.y]
+    linestyleflag = aes.linestyle ≠ nothing
+    colorflag = aes.color ≠ nothing
+    aes_color = colorflag ? aes.color : [nothing] 
+    aes_group = (aes.group ≠ nothing) ? aes.group : [nothing]
+    CT, GT = eltype(aes_color), eltype(aes_group)
+
+    groups = collect(Tuple{CT, GT}, Compose.cyclezip(aes_color, aes_group))
+    ugroups = unique(groups)
+    nugroups = length(ugroups)
+
+    K, V = Tuple{CT, GT}, Matrix{eltype(Dat)}
+    
+    grouped_xy = if nugroups==1
+        Dict{K, V}(ugroups[1]=>Dat)
+    elseif nugroups>1
+        Dict{K, V}(g=>Dat[groups.==[g],:] for g in ugroups)
+    end
+
+    aess = Gadfly.Aesthetics[]
+    for (g, data) in grouped_xy
+        aes1 = Gadfly.Aesthetics()
+        k = KernelDensity.kde(data, bandwidth=window, npoints=stat.n)
+        aes1.x, aes1.y, aes1.z = k.x, k.y, k.density
+        apply_statistic(ContourStatistic(levels=stat.levels), scales, coord, aes1)
+        colorflag && (aes1.color = fill(g[1], length(aes1.x)))
+        push!(aess, aes1)
+    end
+    
+    aes2 = Gadfly.concat(aess...)
+    aes.x, aes.y, aes.color = aes2.x, aes2.y, aes2.color
+    colorflag && (linestyleflag ? (aes.group = aes2.group) : (aes.linestyle = aes2.group.+1))
+    if !colorflag
+        aes.group = aes2.group
+        aes.color_function = aes2.color_function
+        aes.color_label = aes2.color_label
+        aes.color_key_colors = aes2.color_key_colors
+        aes.color_key_continuous = aes2.color_key_continuous
+    end
+    
 end
 
 
@@ -2124,9 +2159,83 @@ function apply_statistic(stat::DodgeStatistic,
     end
 end
 
+struct QuantileBarsStatistic <: Gadfly.StatisticElement
+    quantiles::Vector{Float64}
+    # We cannot avoid these by combining our statistic with Stat.density,
+    # because we need the raw data as well as the kernel density.
+    n::Int # Number of points sampled.
+    bw::Real # Bandwidth used for the kernel density estimation.
+end
+QuantileBarsStatistic(; quantiles=[0.025, 0.975], n=256, bandwidth=-Inf) =
+    QuantileBarsStatistic(quantiles, n, bandwidth)
 
+input_aesthetics(stat::QuantileBarsStatistic) = [:x]
+output_aesthetics(stat::QuantileBarsStatistic) = [:x, :y, :xend, :yend]
 
+"""
+    Stat.quantile_bars[(; quantiles=[0.025, 0.975], bar_width=0.1, n=256, bandwidth=-Inf)]
 
+Transform the point in $(aes2str(input_aesthetics(quantile_bars()))) into a set of
+$(aes2str(output_aesthetics(quantile_bars()))) points. These points can then be drawn
+via [`Geom.segment`](@ref Gadfly.Geom.segment). Here, `bandwidth` works independently
+from the `bandwidth` setting for `Stat.density`.
+"""
+const quantile_bars = QuantileBarsStatistic
 
+"""
+    _calculate_quantile_bar(stat::QuantileBarsStatistic, aes)
+
+Helper function for `apply_statistic(stat::QuantileBarsStatistic, ...)`.
+"""
+function _calculate_quantile_bar(stat::QuantileBarsStatistic, xs)
+    isa(xs[1], Real) || error("Kernel density estimation only works on Real types.")
+
+    window = stat.bw <= 0.0 ? KernelDensity.default_bandwidth(xs) : stat.bw
+    k = KernelDensity.kde(xs, bandwidth=window, npoints=stat.n)
+
+    x = quantile(xs, stat.quantiles)
+    y = zeros(length(x))
+    xend = x
+    yend = pdf(k, x)
+
+    return x, y, xend, yend
+end
+
+function apply_statistic(stat::QuantileBarsStatistic,
+                       scales::Dict{Symbol, Gadfly.ScaleElement},
+                       coord::Gadfly.CoordinateElement,
+                       aes::Gadfly.Aesthetics)
+    Gadfly.assert_aesthetics_defined("QuantileBarsStatistic", aes, :x)
+
+    if aes.color === nothing
+        aes.x, aes.y, aes.xend, aes.yend = _calculate_quantile_bar(stat, aes.x)
+    else
+        groups = Dict()
+        for (x, c) in zip(aes.x, Gadfly.cycle(aes.color))
+            if !haskey(groups, c)
+                groups[c] = Float64[x]
+            else
+                push!(groups[c], x)
+            end
+        end
+
+        colors = Array{Gadfly.RGB{Float32}}(undef, 0)
+        aes.x = Array{Float64}(undef, 0)
+        aes.y = Array{Float64}(undef, 0)
+        aes.xend = Array{Float64}(undef, 0)
+        aes.yend = Array{Float64}(undef, 0)
+        for (c, xs) in groups
+            x, y, xend, yend = _calculate_quantile_bar(stat, xs)
+
+            append!(aes.x, x)
+            append!(aes.y, y)
+            append!(aes.xend, xend)
+            append!(aes.yend, yend)
+            append!(colors, fill(c, length(x)))
+        end
+        aes.color = discretize_make_ia(colors)
+    end
+    aes.y_label = Gadfly.Scale.identity_formatter
+end
 
 end # module Stat